requirements.txt
torch
torchvision
transformers
scikit-learn
pandas
Pillow
tqdm
